pyproject.toml
[tool.poetry]
name = "python_profiler"
version = "0.1.0"
description = ""
authors = ["Azeem <azmaktr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
